Exposing header/canvas components
